Cleanup, check bracket for timestamp
In this version a fix had to ensure lines of the pod description begins with an opening bracket. Before additional texts such as links to this site here lead to generation of fake chapters. One could say a trap of its own success.
Many lines of the initial script were deleted since they were left from a template script of a completely different origin.